examples/voice_agent_requirements.txt
websockets
scipy
sounddevice
numpy
python-dotenv
